Impoved the project directory organization.
Intoroduce new controllers for each view
Created a folder for each view to all associated views.
