+++ code/TicketApp/TicketAppWeb/Controllers/GroupController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace TicketAppWeb.Controllers
+{
+    public class GroupController : Controller
+    {
+        [HttpGet]
+        public IActionResult GroupManagement() 
+        { 
+            return View(); 
+        }
+    }
+}
