+++ code/TicketApp/TicketAppWeb/Controllers/ProjectContorller.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace TicketAppWeb.Controllers
+{
+    public class ProjectController : Controller
+    {
+        [HttpGet]
+        public IActionResult ProjectManagement()
+        {
+            return View();
+        }
+    }
+}
