+++ code/TicketApp/TicketAppWeb/Controllers/UserController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace TicketAppWeb.Controllers
+{
+    public class UserController : Controller
+    {
+        [HttpGet]
+        public IActionResult UserManagement()
+        {
+            return View();
+        }
+    }
+}
