+++ code/TicketApp/TicketAppWeb/TicketAppWeb.csproj
@@ -25,4 +25,8 @@
     <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="9.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\Configuration\" />
+  </ItemGroup>
+
 </Project>
