+++ code/TicketApp/TicketAppWeb/Views/Shared/_Layout.cshtml
@@ -1,4 +1,4 @@
-﻿@using TicketAppWeb.Models;
+﻿@using TicketAppWeb.Models.DomainModels;
 @{
     var currentController = ViewContext.RouteData.Values["controller"]?.ToString() ?? "";
     var currentAction = ViewContext.RouteData.Values["action"]?.ToString() ?? "";
@@ -31,21 +31,21 @@
             </li>
             <li>
                 <a class="nav-link @Nav.Active("UserManagement", currentAction)"
-                   asp-action="UserManagement" asp-controller="Home" asp-area="">
+                   asp-action="UserManagement" asp-controller="User" asp-area="">
                     <i class="fas fa-user"></i>
                     <span class="nav-item">User Management</span>
                 </a>
             </li>
             <li>
                 <a class="nav-link @Nav.Active("GroupManagement", currentAction)"
-                   asp-action="GroupManagement" asp-controller="Home" asp-area="">
+                   asp-action="GroupManagement" asp-controller="Group" asp-area="">
                     <i class="fas fa-users"></i>
                     <span class="nav-item">Group Management</span>
                 </a>
             </li>
             <li>
                 <a class="nav-link @Nav.Active("ProjectManagement", currentAction)"
-                   asp-action="ProjectManagement" asp-controller="Home" asp-area="">
+                   asp-action="ProjectManagement" asp-controller="Project" asp-area="">
                     <i class="fas fa-project-diagram"></i>
                     <span class="nav-item">Project Management</span>
                 </a>
@@ -53,7 +53,7 @@
         </ul>
     </nav>
     <main>
-        @RenderBody();
+        @RenderBody()
     </main>
 </body>
 </html>
